otel: allow filebeat otel to load otel config files (#45093)
* otel: allow filebeat otel to load otel config files

Currently the filebeat otel command only loads beats config files, which
it then translates to beats receivers in order to run a OTel collector.

This PR augments this functionality to allow the command to load otel
configuration files, this aids in testing and running beats receivers in a standalone collector.

* update notice

* add test for multiple receivers

* fix monitoring port

* test refactor for template

* restore condition removed by mistake

* allow monitoring http request with variable url