[Metricbeat] Fix CPU/disk/network visualizations in EC2 overview dashboard (#45117)
Updates the CPU/disk/network metric field names in the "[Metricbeat AWS] EC2 Overview" dashboard.

The "[Metricbeat AWS] EC2 Overview" dashboard that comes with the AWS modules was still using the old metric field names that are no longer in use, resulting in a dashboard with multiple empty visualizations.

Here are the current changes:


| Old                          | New                          | Formatter |
| ---------------------------- | ---------------------------- | --------- |
| `aws.ec2.cpu.total.pct`      | `host.cpu.usage`             | `percent` |
| `aws.ec2.diskio.read.bytes`  | `host.disk.read.bytes`       | `bytes`   |
| `aws.ec2.diskio.write.bytes` | `host.disk.write.bytes`      | `bytes`   |
| `aws.ec2.network.in.bytes`   | `host.network.ingress.bytes` | `bytes`   |
| `aws.ec2.network.out.bytes`  | `host.network.egress.bytes`  | `bytes`   |

---------

Co-authored-by: kaiyan-sheng <kaiyan.sheng@elastic.co>