[Filebeat/Filestream] Delete ingested log files on EOF or inactive (#43368)
This commit allows the Filestream input to delete the log files it is ingesting when the reader is closed and
all events from the file have been acknowledged.

First the reader needs to be closed (either due to EOF or inactivity), then the harvester checks if all events have been published, if so, then it waits for the grace period, checks if no data has been added to the file and finally removes it.

If any of those checks fail or removing the file fails, then the harvester is closed and it will be re-started in the next scan.

A new event type, `OpNotChanged` has been added to the filewatcher, this event is used to re-start the harvester for files that should be removed but have not changed.

The new configuration options and their defaults are:
 - `delete.enabled: false`
 - `delete.grace_period: 30m`

Some log entries from the harvester life-cycle have been updated to better express what Filestream is doing.
---------

Co-authored-by: Craig MacKenzie <craig.mackenzie@elastic.co>
Co-authored-by: Fabrizio Ferri-Benedetti <fabri.ferribenedetti@elastic.co>