x-pack/filebeat/input/httpjson: allow empty cursor template values to be ignored by health status updates
Health status changes are evaluated in CI for integrations and in some
cases (aws.guardduty is the provocation here), there is a failed/empty
evaluation. So allow ignore_empty_value configurations for cursor
entries to also be ignored for health status updates, instead of being
noted as a health degradation.