Handle overflows
We now check for G115, most overflows are impossible, like converting s63
seconds to u64 seconds for date (will overflow in 292 billion years).

Pids are actually 32bit in the kernel so casting * -> u32 is safe.

This is a backport, and I'd hate to introduce a bug by adding unecessarily
overflow handling.