x-pack/filebeat/input/httpjson: share oauth2 clients where possible (#44976)
The httpjson input makes new http.Client values for every chain step.
This is inefficient for a variety of reasons, but the configuration API
that we expose makes the general fix for this non-trivial. It is,
however, possible to fix the OAuth2 behaviour that arises from this
where root parent steps in a chain make use of OAuth2 and so do its
children, but the children needlessly request tokens because they do
not have access to the oauth2.Transports token cache. The fix here is
to retain the prepared OAuth2 client in the relevant config and use
that if it is non-nil. This does not fix the case that a non-root
parent shares OAuth2 authentication with its children, but this is
likely to be uncommon.

(cherry picked from commit 81d304e929c7c5a1802cb1be65c63e30cb9a2314)