[filebeat] add experimental GZIP file ingestion in filestream (#45301)
This change introduces experimental support for ingesting GZIP-compressed
files using the filestream input. To use this functionality, set the new
`gzip_experimental` configuration option to true.

GZIP files are read in their entirety and then closed. To prevent
re-ingestion after a restart, the final message from the file is
flagged with an EOF marker. This marker is saved to the registry,
instructing Filebeat not to process the file again.

The implementation uses a new `EOFLookaheadReader` to identify the
final message and assign `io.EOF` to its `Private` field.

To support monitoring, filestream metrics are updated. A `gzip_`
variant has been added for each relevant metric to track compressed
files exclusively. The pre-existing metrics now report aggregate
values for all file types (plain and GZIP).