docs/reference/filebeat/cel: use heavier applies_to tag for headings
Unfortunately, in-line tags cannot be used in heading since the ToC generator
attempts to render them in the table, and fails. Ideally, this would work since
the in-line approach produces a lighter documentation. Revert this change in
future if doc engineering can be convinced of this and heading handling is fixed
for this approach.