x-pack/filebeat/input/cel: clarify logging for periodic starts and budget excession
The periodic state log message "process repeated request" could be
interpreted to mean that a request is being re-sent, so replace this
with "process periodic request".

Although the budget excession logging is only a warning, it could be
interpreted to signify that an error state has happened, possibly
leading to data loss. Add text to indicate that normal operation will
continue and give an indication of when. Note that the next_eval_time is
an estimate since the actual time is under the control of a time.Ticker
in the timed.Periodic function, and tickers may adjust arrival of
ticks to make up for slow receivers[1]. This should not be an issue for
configurations unless the amount of time spent in an evaluation is on
the order of the configured interval.

[1]https://pkg.go.dev/time#Ticker