Fix hints default config for logs to match Kubernetes and Docker
The hints default config template was using
`${data.kubernetes.container.id}`, however it only works with the
Kubernetes provider.

To also support the Docker provided the template is updated to use
`${data.container.id}` that is supported by both providers.

The paths is also updated to include the Kubernetes and Docker log
paths.