[elastic-agent-system-metrics] feat: deprecate SetupMetrics and migrate to SetupMetricsOptions (#258)
## What does this PR do?

instead of breasking the method signature move to arg options

## Why is it important?

to account for consumers and older branches

## Checklist

<!-- Mandatory
Add a checklist of things that are required to be reviewed in order to
have the PR approved

List here all the items you have verified BEFORE sending this PR. Please
DO NOT remove any item, striking through those that do not apply. (Just
in case, strikethrough uses two tildes. ~~Scratch this.~~)
-->

- [ ] My code follows the style guidelines of this project
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added tests that prove my fix is effective or that my
feature works
- [ ] I have added an entry in `CHANGELOG.md`

## Author's Checklist

<!-- Recommended
Add a checklist of things that are required to be reviewed in order to
have the PR approved
-->
- [ ]

## Related issues

<!-- Recommended
Link related issues below. Insert the issue link or reference after the
word "Closes" if merging this should automatically close it.

- Closes #123
- Relates #123
- Requires #123
- Superseds #123
-->
-

---------

Co-authored-by: Khushi Jain <khushi.jain@elastic.co>
Co-authored-by: Vihas Makwana <121151420+VihasMakwana@users.noreply.github.com>