[AutoOps] Provide flag to see Enterprise License in 7.x
This updates the code to dynamically specify the
`?acknowledge_enterprise` query string parameter so that it returns
`enterprise` if the license is enterprise in 7.x (without the
parameter, 7.x will return `platinum` for compatibility reasons).

This flag is unneeded in all other versions, so it is only provided
in 7.x.

It also adds a lot of tests for the maybeRegister code to test the
different scenarios.

Finally, as a secondary benefit, this logs the registered ID when
CCM is successfully registered with the Cluster ID (one time!).