[PoC] Reading another registry and taking over from its log input
That's just to prove that we can open a different store while
initialising the Filestream input and take over any states from
it. While the code works, there are some hard coded paths as well as
there are no safety checks or proper error handling.

Aside from that, it works well.