x-pack/httpjson: improve/fix logic for health status updates
There was no way to ask the input to not report degraded health in the
case that a template error was required to signal that some aspect of
the input's operation should terminate, outside of the template itself.
This change fixes that.

Also fix buggy cursor template tests that were identified by this
improvement.