[AutoOps] Allow `AUTOOPS_ES_` prefix for environment variables (#46358)
This adds support for three expected environment variables to be
supplied by service usage of AutoOps wrapping the Elasticsearch
metricset.

This adds support for these in addition to the existing key

- `AUTOOPS_ES_USERNAME` (or `ELASTICSEARCH_READ_USERNAME`)
- `AUTOOPS_ES_PASSWORD` (or `ELASTICSEARCH_READ_PASSWORD`)
- `AUTOOPS_ES_API_KEY` (or `ELASTICSEARCH_API_KEY`)

This also adds support for the API Key to be supplied as a Base64
encoded value, which will be used directly. This is detected by
looking for the `:` that separates the existing `id:key` value,
then encoding it if it detects that and, otherwise, using it
directly.

(cherry picked from commit b0f8ee7c69afae2250d694eb1228d483c4eff7c3)