dgram.Listener return error on connection failure (#46302)
The Listener.doRun method previously logged connection failures at a
debug level but did not return an error to the caller. This made it
difficult for users to understand why a input was not working.

This commit modifies Listener.doRun to return an error when a
connection failure occurs and also updates the UDP and Unix inputs to
set their status as failed when an error is returned by Listener.Run.

* Add changelog

* Fix log level

* mage check

* double test timeout

* Try using 127.0.0.1 instead of all interfaces

* TestLogStatusReporter run Filebeat as a sub process

TestLogStatusReporter was starting the Filebeat
command (github.com/elastic/beats/v7/x-pack/filebeat/cmd.Filebeat())
as part of the test, this loaded the default seccomp policy, blocking
system calls needed by subsequent tests like 'pidfd_send_signal'.

This commit fixes the problem by running Filebeat as a sub process,
like all the other integration tests, thus providing the necessary
isolation between tests.