[Flaky Test] Tentative fix for fbreceiver and filestream flaky tests (#46531)
* tentative fix for TestReceiverDegraded, TestConsumeContract and TestFilestreamTruncateBlockedOutput

These tests are currently marked as flaky.

TestReceiverDegraded and TestConsumeContract are failing on linux/arm64
due to goroutine leaks from TCP clients.

TestFilestreamTruncateBlockedOutput is likely failing due to a small
assertion time in env.requireOffsetInRegistry.

For the goroutine leaks, allow the TCP client goroutines, since they are
hard to track and could be from other tests. The stacktrace only goes as
far as the standard library os package.

For the filestream test, increase the timeout for the assertion.

* skip TestConsumeContract