Filestream validate `clean_inactive` and disable when 0 (#46373)
There was no validation of the clean_inactive setting even though the
documentation is very clear on the restrictions about using it.

This commit adds the necessary validations causing Filebeat to exit if
they fail.

Setting `clean_inactive: 0` now disables the option instead of the
buggy behaviour of re-ingesting all files on restart.

To maintain backwards compatibility, `legacy_clean_inactive` can be set to
`true`, restoring the previous behaviour of not validating `clean_inactive` and
re-ingest files on restart.

---------

Co-authored-by: Anderson Queiroz <me@andersonq.me>