[libbeat/output/kafka]: fix panic on Publish after Close (#46446)
* libbeat/output/kafka: fix panic on Publish after Close

A race condition during client shutdown could cause a panic.

If a `Publish` call was in-flight while the client was closing, the
underlying Kafka producer's input channel could be closed before the
call finished sending all messages in the batch. The subsequent attempt
to send the remaining messages on this closed channel would result in a
panic.