[Filebeat/Journald] Add support for watching folders (#46682)
This commit implements the use of `--directory` argument when calling
`journalctl`, thus allowing it to watch for new files in the folder.

This is implemented by calling `os.Stat` on the path and checking if
it is a folder, if so, then `--directory` is used instead of `--file`.

(cherry picked from commit 8af95b5d17f3ea68b73db69564682a450e508ad4)