Use noop logger instead of logger backed by t.Log
It can happen that the manager only finishes to shutdown after the
test has finished, that is somehow expected because ManagerV2.Stop
does not block. However this can cause a panic if the logger passed to
the manager is backed by t.Log.

This commit fixes it by using a noop logger.