[Filebeat] set `close.on_state_change.removed: false` in hints default config (#46695)
When using autodiscover it the last log lines of a file might not be
not fully ingested, this happens because the file is removed shortly
after the container is removed, not giving Filebeat enough time to
ingest the full file.

This commit mitigates the problem by setting
`close.on_state_change.removed: false` in the hints default
configuration and adding it to all configuration examples.

A new page in the troubleshooting section is also added.

---------

Co-authored-by: Visha Angelova <91186315+vishaangelova@users.noreply.github.com>