[Filbeat/Journald] Restore support for paths like /foo/*/* (#46792)
When the --directory option was introduced we broke the behaviour of
supporting globs like /foo/*/*, this commit restores it.

If there is any error when trying to stat the path to check whether
it is directory, we log at debug level and default to the previous
behaviour of using `--path`.