x-pack/filebeat/input/cel: add support for dynamic rate limit setting (#46953)
This adds support for directly setting the mito runtime's limiter's limit
and burst before return from the CEL evaluation. This is important for
cases where the CEL program is making multiple requests to a set of
related endpoints that share a rate limit budget before the CEL program
returns as it enables the program/runtime to conform to the API's limit
policies without returning from the CEL program.