[elastic-agent-system-metrics] Remove "Accurate CPU counts not available on platform" log spam on MacOS at the debug level (#266)
See https://github.com/elastic/beats/issues/47054. The removed debug log
generated 10s to 100s of lines of log spam in Metricbeat and Elastic
Agent logs at the debug level on my MacOS machine. We can see that
`numCPU` isn't implemented on a platform by inspection in the code, we
don't need a log for it.

Also for some reason the notice file was out of date, so that is updated
too.