[Filebeat/Filestream] Prevent modified files close due to inactivity
This commit prevents files to be closed due to inactivity if their
size has changed.

A race condition could happen when the file watcher would send a write
event, the harvester was still running, so a new one would not start,
however at about the same time, the inactive timeout would be reached,
closing the harvester before the data that triggered the write event
was ingested.

This commit fixes it by checking if the file size has changed before
closing the file due to inactivity.