fix: add_cloud_metadata: Do not block on String() (#47058)
In `addCloudMetadata`, do not block by waiting for `init()` when calling `String()`, but return `add_cloud_metadata=<uninitialized>`. After the timeout, 3 seconds by default, `add_cloud_metadata` will be logged as previously.

I've decided to keep the old log format because this is also dumped during agent diagnostics and I figured the complexity is not significant. We can still go for a simpler approach of not logging any metadata instead.