[WIP] Run Log input system tests with Filestream
This is the beginning of running the system-tests (Python tests) from
the Log input as Filestream input.

The feature flag `log_input_run_as_filestream` is enabled in the
configuration files when the environment variable `RUN_AS_FILESTREAM`
is true.

`filebeat/tests/system/test_json.py` is updated to work with both
inputs.