libbeat/processors/translate_ldap_attribute: fix double locking and improve logging
Previously, when findObjectBy was called and the connection was closing,
the reconnect method would both hold the lock and call connect which
attempts to then gain the lock, resulting in a deadlock. Reduce the
level of locking; I believe this can be reduced further, but the
concurrent use protections here are unclear, so I have not gone further.

Also improve debug logging during dial and query.