[azureblobstorage] - Make OAuth2 tests compatible with latest version of AzIdentity and AzureAD libraries (#47611)
Microsoft in their latest SDK updates to AzureAD and And AzIdentity, have
made changes to how their OIDC endpoints are validated. With these changes, when
using a custom source for the OIDC endpoints, the "issuer" must be validated
against the "authority" (base url). With these changes when we try to inject a
custom authority in the http transport, completely separate clients seem to be
created internally by the SDK which bypass our custom transport and we are unable
to mock the authentication process. Because of this, we are changing the mock
authentication to occur on standard Microsoft endpoints which bypass the authority
validation as a whole.

(cherry picked from commit d333f3c19410b2f6802e4c852190ad5d03a242c4)