chore(otel): remove otelbeat (#47804)
The otelbeat binary is not used anymore since we moved to an in-process testing collector in #47338, so remove the subcommand and all references to it in the build system. The otel subcommand was gated behind a build tag and only used in tests.

(cherry picked from commit f4b9ad1dd6d222c3fcbf7e154a22593d67177b8f)

# Conflicts:
#	x-pack/filebeat/magefile.go
#	x-pack/libbeat/common/otelbeat/otel.go