test(TestDurationIsAddedToEvent): wait for worker to shutdown (#47816)
metricbeat wrapper is leaking a goroutine causing the test to fail
because it's logging a message after the test is over.

Wait for worker to shutdown before returning.

Similar to #47660 which fixed the same issue for TestPeriodIsAddedToEvent.
Similar to #47663 which fixed the same issue for TestGenerateProcessorList.

Closes #47776

(cherry picked from commit a880dc2575ac2d8617f77ecb63161d99e7634afe)