[AutoOps] Show Better Error Messages (#48204)
This fixes an issue where we did not report the actual error message
and therefore obscured actual issues.

Along with this change, and benefiting from it, we now analyze the
response for the Cloud Connect API to see if it has an error and
push that error downstream.

(cherry picked from commit fe5c564e38103ea1caed9a1bc307916abeedf884)